+++ src/main/kotlin/com/github/qianmi/action/shell/ShellSelectedAction.kt
@@ -1,0 +1,19 @@
+package com.github.qianmi.action.shell
+
+import com.github.qianmi.domain.project.tools.ShellElement
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+
+
+abstract class ShellSelectedAction : AnAction() {
+    abstract var ele: ShellElement
+
+    /**
+     * 开启
+     */
+    abstract fun open(e: AnActionEvent)
+
+    override fun actionPerformed(e: AnActionEvent) {
+        this.open(e)
+    }
+}