+++ src/main/kotlin/com/github/qianmi/action/shell/ssh/SshSelectedAction.kt
@@ -1,13 +1,33 @@
 package com.github.qianmi.action.shell.ssh
 
-import com.github.qianmi.domain.project.tools.Shell
-import com.intellij.openapi.actionSystem.AnAction
+import com.github.qianmi.action.shell.ShellSelectedAction
+import com.github.qianmi.domain.project.AllProject
+import com.github.qianmi.domain.project.tools.ShellElement
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.Project
+import com.jetbrains.plugins.remotesdk.console.SshConsoleOptionsProvider
+import com.jetbrains.plugins.remotesdk.console.SshTerminalCachingRunner
+import org.jetbrains.plugins.terminal.TerminalTabState
+import org.jetbrains.plugins.terminal.TerminalView
 
 
-class SshSelectedAction(var ele: Shell.Element) : AnAction() {
+class SshSelectedAction(override var ele: ShellElement) : ShellSelectedAction() {
 
-    override fun actionPerformed(e: AnActionEvent) {
-        ele.openSshTerminal(e.project!!)
+    override fun open(e: AnActionEvent) {
+        val project = e.project!!
+
+        val charset = SshConsoleOptionsProvider.getInstance(project).charset
+        val runner = SshTerminalCachingRunner(project, ele.createSshUiData(), charset)
+        //开启openssh
+        TerminalView.getInstance(project).createNewSession(runner, buildTerminalTabState(project))
+    }
+
+    private fun buildTerminalTabState(project: Project): TerminalTabState {
+        val myProject = AllProject.currentProject(project)
+
+        val state = TerminalTabState()
+        state.myTabName = "${myProject.env.envName}(${ele.group})：${ele.ip}"
+        state.myWorkingDirectory = ele.getWorkingDirectory()
+        return state
     }
 }